even more component reorganization!
